ref: [Bug][NII Redmine#36946]エクスポート関連のラベル変更: - Add dialog confirm when delete export storage storage location - Fix JP label

+++ admin/static/js/rdm_custom_storage_location/rdm-institutional-storage-page.js
@@ -751,8 +751,17 @@ afterRequest.delete = {
 $('.delete-location').click(function () {
     var id = this.dataset.id;
     // var providerShortName = this.dataset.provide;
-
-    deleteLocation(id);
+      $osf.confirmDangerousAction({
+        title: _('Are you sure you want to delete this export data storage location?'),
+        callback: function() {
+            deleteLocation(id);
+        },
+        buttons: {
+            success: {
+                label: _('Delete')
+            }
+        }
+    });
 });
 
 function deleteLocation(id) {
